SLEMA-79 Moved data related files to a different directory
