+++ SlemaForAndroid/lib/features/well_being_assessment/data/symptom_type.dart
