+++ SlemaForAndroid/lib/features/well_being_assessment/data/well_being_variant.dart
