+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_symptom_entry.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_subtitle_text.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_symptoms_change_value_button.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/symptom_type.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/symptom_value.dart';
+import 'package:pg_slema/features/well_being_assessment/data/symptom_type.dart';
+import 'package:pg_slema/features/well_being_assessment/data/symptom_value.dart';
 
 class AssessmentSymptomEntry extends StatelessWidget {
   const AssessmentSymptomEntry(
