+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_symptoms_form.dart
@@ -3,8 +3,8 @@ import 'package:pg_slema/features/well_being_assessment/presentation/widget/form
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_form_title.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_manage_symptoms_button.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_symptom_entry.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/symptom_type.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/symptom_value.dart';
+import 'package:pg_slema/features/well_being_assessment/data/symptom_type.dart';
+import 'package:pg_slema/features/well_being_assessment/data/symptom_value.dart';
 import 'package:pg_slema/utils/log/logger_mixin.dart';
 
 class AssessmentSymptomsFormWidget extends StatelessWidget with Logger {
