+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/widget/forms/well_being/assessment_well_being_selector.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/well_being/assessment_well_being_selector_variant.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/well_being/well_being_variant.dart';
+import 'package:pg_slema/features/well_being_assessment/data/well_being_variant.dart';
 import 'package:pg_slema/utils/log/logger_mixin.dart';
 
 class AssessmentWellBeingSelector extends StatelessWidget with Logger {
